feat: add extra validation
